+++ src/util/enumerateDiagnostics.ts
@@ -3,6 +3,8 @@ import LLAlignAnd from '../LL/LLAlignAnd';
 import LLAlignEnd from '../LL/LLAlignEnd';
 import LLAlignSingleLine from '../LL/LLAlignSingleLine';
 import LLBig from '../LL/LLBig';
+import LLBracketCurly from '../LL/LLBracketCurly';
+import LLBracketRound from '../LL/LLBracketRound';
 import LLColonEqq from '../LL/LLColonEqq';
 import LLColonForMapping from '../LL/LLColonForMapping';
 import LLCref from '../LL/LLCref';
@@ -27,6 +29,8 @@ export default function enumerateDiagnostics(doc: vscode.TextDocument): vscode.D
         LLAlignEnd,
         LLAlignSingleLine,
         LLBig,
+        LLBracketCurly,
+        LLBracketRound,
         LLColonEqq,
         LLColonForMapping,
         LLCref,